Add proper reverse-proxy.
I was trying to get websockets to work over stormgrok and learned
quite a bunch about reverse proxies.. There is particular headers
called the hop-by-hop headers which should be stripped. Connection
upgrade requests should _not_ be forwarded to clients unless it's
a websocket upgrade request or the server thinks it's a good idea..
:shrug: aand the Forwarded header should be populated with data.

This seemed like a lot work and I found a package that does it..
Problem is that it's old, poorly maintained, ws support came in
recently but is not yet on crates it uses the X-Forwaded header
instead of the proper Forwaded header and its 350 lines instead of
the 10 I currently use.. But, ya know, it does actually work.